Feature/order item service controller (#7)
* feat : 주문 아이템 dto 클래스 추가

* feat : 주문 아이템 crud 서비스 추가

* feat : 주문 아이템 엔티티에서 업데이트 기능 추가

* feat : 주문 저장시에 유저에서 저장하도록 코드 수정함

* test : 주문 아이템 저장 테스트 코드 작성

* refactor : price, quantity를 Long 타입으로 변경

* refactor : dto 클래스에 상품 이름 추가

* feat : 다른 엔티티의 아이디로 디비에서 가져올 때 List형태로 가져오도록 코드 수정함

* feat : order item 조회, 등록, 등록 페이지 컨트롤러 작성

* feat : order item 저장 ajax 요청 기능 추가

* feat : order를 유저로 조회할 때 list로 가져오도록 코드 수정함

* feat : 주문 아이템 저장 페이지 생성

* feat : 주문 아이템 상세 페이지 구현, 주문 아이템 삭제 기능 구현

* feat : 주문 삭제시 상품 개수에 주문 아이템 개수 더하는 기능 추가

* feat : 주문 아이템 상세 페이지 연동

* feat : 주문 아이템 삭제 기능 추가

* feat : 주문 아이템 상세 페이지 추가

* feat : 유저 삭제시 주문 아이템 개수 상품에 추가하도록 기능 구현

* feat : 페이지간 이동 기능 추가

+++ src/main/java/kr/ac/cnuswacademy/springcoffeeshop/dto/product/ProductResponseDto.java
@@ -7,8 +7,8 @@ import lombok.Getter;
 public class ProductResponseDto {
     private Long id;
     private String name;
-    private int price;
-    private int quantity;
+    private Long price;
+    private Long quantity;
     private String description;
 
     public ProductResponseDto (Product product) {
